refactor(hooks): extract useUserIdResolution from useContactForm
useContactForm mixed three concerns across 520 lines. The user-ID
resolution pipeline (format validation, MNS resolution, public-key
fetch/cache, already-a-contact check, stale-response sequence guard)
now lives in useUserIdResolution; the previously duplicated MNS and
plain-ID tails share one fetchAndValidateId helper. useContactForm's
public surface is unchanged (NewContact untouched).
